requirements.txt
Django==3.2.5
djangorestframework==3.12.4
graphene==3.3
graphene_django==3.1.5
djangorestframework-simplejwt==4.7.2
